backend/requirements.txt
faiss_cpu==1.10.0
fastapi==0.115.9
langchain_community==0.3.18
langchain_core==0.3.40
langchain_huggingface==0.1.2
langchain_text_splitters==0.3.6
numpy==2.2.3
PyPDF2==3.0.1
python-dotenv==1.0.1
python_docx==1.1.2
rank_bm25==0.2.2
SQLAlchemy==2.0.38
torch==2.6.0
uvicorn==0.34.0
llama-cpp-python==0.3.5
python-multipart==0.0.20
websockets==15.0
uvicorn==0.34.0